moved commands, added use
